add reminder

+++ gotcha/app/src/main/java/volunteer/sk/greate43/com/gotcha/Reminder.java
@@ -3,8 +3,8 @@ package volunteer.sk.greate43.com.gotcha;
 public class Reminder {
     private String reminderId;
     private String reminderName;
-    private long lat;
-    private long lon;
+    private double lat;
+    private double lon;
 
     public Reminder() {
     }
@@ -25,19 +25,19 @@ public class Reminder {
         this.reminderName = reminderName;
     }
 
-    public long getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(long lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public long getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(long lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 }
